requirements.txt
Django==1.8.2
Markdown==2.6.2
Pillow==2.8.2
PyJWT==1.4.0
South==1.0.2
Unidecode==0.04.18
argparse==1.2.1
boto==2.39.0
defusedxml==0.4.1
django-autocomplete-light==2.2.2
django-autoslug==1.8.0
django-braces==1.8.1
django-cities==0.4.1
django-cities-light==3.1.2
django-cors-headers==1.1.0
django-filter==0.10.0
django-oauth-toolkit==0.8.2
django-paypal==0.2.5
django-redis-sessions==0.5.0
django-rest-framework-social-oauth2==0.0.8
django-ses==0.7.0
django-smtp-ssl==1.0
django-social-auth==0.7.28
django-suit==0.2.15
django-websocket-redis==0.4.4
django-wysiwyg-redactor==0.4.9
djangorestframework==3.1.2
djangotoolbox==1.8.0
gevent==1.0.2
greenlet==0.4.9
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.0.0
parse==1.6.6
psycopg2==2.6.1
python-openid==2.2.5
python-social-auth==0.2.12
python3-openid==3.0.5
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.10.0
sorl-thumbnail==12.3
urllib3==1.14
wsgiref==0.1.2
